Added UIController class and moved all controller classes into their own folder
