+++ Program.cs
@@ -1,4 +1,4 @@
-﻿using Transactor;
+﻿using Transactor.controllers;
 
 if (args.Length != 1)
 {
@@ -9,18 +9,4 @@ if (args.Length != 1)
 FileController fileController = new();
 fileController.Read(args[0]);
 
-string? userInput = "";
-while (userInput != "exit")
-{
-    Console.WriteLine("How would you like to view your transactions? Options: [list] | [exit] to exit the program");
-    userInput = Console.ReadLine();
-
-    switch (userInput)
-    {
-        case "list":
-            Console.WriteLine("Listing your transactions...");
-            break;
-        default:
-            break;
-    }
-}
+UIController.Start();
