+++ controllers/FileController.cs
@@ -1,4 +1,4 @@
-namespace Transactor;
+namespace Transactor.controllers;
 
 public class FileController
 {
